made powered spawner work
for now i have decided to go with an even simpler solution of the "SoulCore solution", and just allow the player to put a filled soul vial into the powered spawner, and that mob will be spawned. My first idea was to use a soul fuser/infuser to combine a filled soul vial and a powered spawner, to give the powered spawner that mob type. Might come back to that idea later, which is why i have pushed the work i made on that as well